+++ lib/src/model_viewer.dart
@@ -2,10 +2,11 @@
 
 import 'dart:async';
 import 'dart:convert';
-import 'dart:io' show Platform;
+import 'dart:io'
+    show HttpRequest, HttpServer, HttpStatus, InternetAddress, Platform;
 
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart' show AssetBundle;
+import 'package:flutter/services.dart' show rootBundle;
 import 'package:flutter_android/android_content.dart' as android_content;
 import 'package:webview_flutter/platform_interface.dart';
 import 'package:webview_flutter/webview_flutter.dart';
@@ -26,7 +27,8 @@ class ModelViewer extends StatefulWidget {
       this.autoRotateDelay,
       this.autoPlay,
       this.cameraControls,
-      this.iosSrc})
+      this.iosSrc,
+      this.proxy = true})
       : super(key: key);
 
   /// The background color for the model viewer.
@@ -73,6 +75,8 @@ class ModelViewer extends StatefulWidget {
   /// via AR Quick Look.
   final String iosSrc;
 
+  final bool proxy;
+
   @override
   State<ModelViewer> createState() => _ModelViewerState();
 }
@@ -81,6 +85,25 @@ class _ModelViewerState extends State<ModelViewer> {
   final Completer<WebViewController> _controller =
       Completer<WebViewController>();
 
+  HttpServer _proxy;
+
+  @override
+  void initState() {
+    super.initState();
+    if (widget.proxy) {
+      _initProxy();
+    }
+  }
+
+  @override
+  void dispose() {
+    super.dispose();
+    if (_proxy != null) {
+      _proxy.close(force: true);
+      _proxy = null;
+    }
+  }
+
   @override
   Widget build(final BuildContext context) {
     return WebView(
@@ -88,19 +111,27 @@ class _ModelViewerState extends State<ModelViewer> {
       javascriptMode: JavascriptMode.unrestricted,
       initialMediaPlaybackPolicy: AutoMediaPlaybackPolicy.always_allow,
       onWebViewCreated: (final WebViewController webViewController) async {
-        print('>>>> ModelViewer initializing...'); // DEBUG
         _controller.complete(webViewController);
-        final AssetBundle bundle = DefaultAssetBundle.of(context);
-        final ThemeData themeData = Theme.of(context);
-        final String htmlTemplate = await bundle
-            .loadString('packages/model_viewer/etc/assets/template.html');
-        final String html = _buildHTML(themeData, htmlTemplate);
-        final String contentBase64 =
-            base64Encode(const Utf8Encoder().convert(html));
-        await webViewController.loadUrl('data:text/html;base64,$contentBase64');
-        final js = await bundle
-            .loadString('packages/model_viewer/etc/assets/model-viewer.js');
-        await webViewController.evaluateJavascript(js);
+        if (widget.proxy) {
+          final host = _proxy.address.address;
+          final port = _proxy.port;
+          print(
+              '>>>> ModelViewer initializing... (proxy at $host:$port)'); // DEBUG
+          await webViewController.loadUrl('http://$host:$port/');
+        } else {
+          print('>>>> ModelViewer initializing...'); // DEBUG
+          final bundle = DefaultAssetBundle.of(context);
+          final themeData = Theme.of(context);
+          final htmlTemplate = await bundle
+              .loadString('packages/model_viewer/etc/assets/template.html');
+          final html = _buildHTML(themeData, htmlTemplate);
+          final contentBase64 = base64Encode(const Utf8Encoder().convert(html));
+          await webViewController
+              .loadUrl('data:text/html;base64,$contentBase64');
+          final js = await bundle
+              .loadString('packages/model_viewer/etc/assets/model-viewer.js');
+          await webViewController.evaluateJavascript(js);
+        }
       },
       navigationDelegate: (final NavigationRequest navigation) async {
         //print('>>>> ModelViewer wants to load: <${navigation.url}>'); // DEBUG
@@ -116,11 +147,10 @@ class _ModelViewerState extends State<ModelViewer> {
             action: "android.intent.action.VIEW", // Intent.ACTION_VIEW
             data: Uri.parse("https://arvr.google.com/scene-viewer/1.0").replace(
               queryParameters: <String, dynamic>{
-                'file': this.widget.src,
+                'file': widget.src,
                 'mode': 'ar_only',
               },
             ),
-            //package: "com.google.android.googlequicksearchbox",
             package: "com.google.ar.core",
             flags: 0x10000000, // Intent.FLAG_ACTIVITY_NEW_TASK,
           );
@@ -147,17 +177,68 @@ class _ModelViewerState extends State<ModelViewer> {
     return HTMLBuilder.build(
       htmlTemplate: htmlTemplate,
       backgroundColor:
-          this.widget.backgroundColor ?? themeData.scaffoldBackgroundColor,
-      src: this.widget.src,
-      alt: this.widget.alt,
-      ar: this.widget.ar,
-      arModes: this.widget.arModes,
-      arScale: this.widget.arScale,
-      autoRotate: this.widget.autoRotate,
-      autoRotateDelay: this.widget.autoRotateDelay,
-      autoPlay: this.widget.autoPlay,
-      cameraControls: this.widget.cameraControls,
-      iosSrc: this.widget.iosSrc,
+          widget.backgroundColor ?? themeData?.scaffoldBackgroundColor,
+      src: widget.proxy ? '/model' : widget.src,
+      alt: widget.alt,
+      ar: widget.ar,
+      arModes: widget.arModes,
+      arScale: widget.arScale,
+      autoRotate: widget.autoRotate,
+      autoRotateDelay: widget.autoRotateDelay,
+      autoPlay: widget.autoPlay,
+      cameraControls: widget.cameraControls,
+      iosSrc: widget.iosSrc,
     );
+  }
+
+  Future<void> _initProxy() async {
+    final url = widget.src;
+    _proxy = await HttpServer.bind(InternetAddress.loopbackIPv4, 0);
+    _proxy.listen((final HttpRequest request) async {
+      print("${request.method} ${request.uri}"); // DEBUG
+      final response = request.response;
+
+      switch (request.uri.path) {
+        case '/':
+          print(request.headers); // DEBUG
+          final htmlTemplate = await rootBundle
+              .loadString('packages/model_viewer/etc/assets/template.html');
+          final html = _buildHTML(null, htmlTemplate) +
+              "<script src=\"/model-viewer.js\"></script>";
+          print(html);
+          response
+            ..statusCode = HttpStatus.ok
+            ..headers.add("Content-Type", "text/html")
+            ..write(html);
+          await response.close();
+          break;
+
+        case '/model-viewer.js':
+          print(request.headers); // DEBUG
+          final code = await rootBundle
+              .loadString('packages/model_viewer/etc/assets/model-viewer.js');
+          response
+            ..statusCode = HttpStatus.ok
+            ..headers
+                .add("Content-Type", "application/javascript;charset=UTF-8")
+            ..write(code);
+          await response.close();
+          break;
+
+        case '/model':
+          print(request.headers); // DEBUG
+          await response.redirect(Uri.parse(url)); // TODO: proxy the resource
+          break;
+
+        case '/favicon.ico':
+        default:
+          response
+            ..statusCode = HttpStatus.notFound
+            ..headers.add("Content-Type", "text/plain")
+            ..write("Resource '${request.uri}' not found");
+          await response.close();
+          break;
+      }
+    });
   }
 }