+++ CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.8.0] - 2020-08-19
+## [0.8.0] - 2020-08-20
+
+### Changed
 
 - Required [flutter_android](https://pub.dev/packages/flutter_android)
   0.8.0+.
+
+### Added
+
+- Support for a `localhost` HTTP server and proxy
 
 ## [0.7.0] - 2020-08-18